Improvements pt. II
