+++ Volume/Models/SettingsPage.swift
@@ -1,0 +1,22 @@
+//
+//  SettingsPage.swift
+//  Volume
+//
+//  Created by Cameron Russell on 4/16/21.
+//  Copyright © 2021 Cornell AppDev. All rights reserved.
+//
+
+import SwiftUI
+
+struct SettingsPage: Identifiable {
+    let id = UUID()
+    
+    let destination: Destination
+    let imageName: String
+    let info: String
+    
+    enum Destination {
+        case externalLink(String)
+        case internalView(NestedSettingsView)
+    }
+}
