+++ Volume/Supporting/Constants.swift
@@ -10,10 +10,7 @@ import Foundation
 
 struct Constants {
     static let missionStatement = """
-    Created by Cornell AppDev, Volume aims to better connect student publications with student readers by making it easy and
-    accessible to digest student created content.\n\nVolume believes in the unique power that lies in the sharing and
-    distribution of content by students and for students to spark conversation, experiences, and connections that improve the
-    collegiate experience.
+    Created by Cornell AppDev, Volume aims to better connect student publications with student readers by making it easy and accessible to digest student created content.\n\nVolume believes in the unique power that lies in the sharing and distribution of content by students and for students to spark conversation, experiences, and connections that improve the collegiate experience.
     """
     static let teamInfo = "The founding members that together made Volume a reality ✨"
